tools/cli/requirements.txt
typer[all]==0.12.3
rich==13.7.1
httpx==0.28.1
pydantic==2.8.2
click==8.1.7
tabulate==0.9.0
python-dotenv==1.0.1
pyyaml==6.0.1
jinja2==3.1.4
pytest==8.3.3
pytest-asyncio==0.23.0
pytest-mock==3.14.0
pytest-cov==5.0.0
